requirements.txt
streamlit
pillow
matplotlib
pytesseract
pandas
python-dotenv
